requirements.txt
streamlit==1.32.0
pandas==2.1.0
plotly==5.18.0
requests==2.31.0
python-dotenv==1.0.0
spotipy==2.23.0
fastapi==0.100.0
uvicorn==0.23.2
scikit-learn==1.3.0
pytest==7.4.0 